+++ src/Commands/Fun/MeepCraft.ts
@@ -3,6 +3,8 @@ import { Message } from "discord.js";
 import { agent } from '../../lib/Utility/Proxy';
 import fetch from 'node-fetch';
 import { Agent } from "https";
+import AbortController from 'node-abort-controller';
+import { AbortSignal } from "node-fetch/externals";
 
 const latest = {
     fetched: 0,
@@ -36,20 +38,31 @@ export default class extends Command {
             return message.channel.send(embed);
         }
         
+        const controller = new AbortController();
+        const timeout = setTimeout(() => controller.abort(), 30000);
+
         let players: { playersOnline: number };
         try {
             const res = await fetch('https://forum.meepcraft.com/game/query.php', {
                 // Love when the only available package for proxying connections
                 // hasn't been properly updated in years. 
                 // https://github.com/TooTallNate/node-https-proxy-agent/issues/108
-                agent: agent as unknown as Agent
+                agent: agent as unknown as Agent,
+                signal: controller.signal as AbortSignal
             });
-
             players = await res.json();
-        } catch {
+        } catch(e) {
+            if(controller.signal.aborted === true) {
+                return message.channel.send(this.Embed.fail(`
+                Request was aborted. Likely caused by the proxy which is being used currently.
+                `));
+            }
+
+            this.logger.log(e);
             return message.channel.send(this.Embed.fail('An unexpected error occurred!'));
         }
 
+        clearTimeout(timeout); // so it doesn't abort after the request succeeded!
         latest.fetched = Date.now();
         latest.results = players.playersOnline;