+++ src/lib/Utility/Rejections.ts
@@ -4,5 +4,6 @@ const logger = new Logger('unhandledRejection');
 
 process.on('unhandledRejection', e => {
     logger.log(e);    
+    console.log(e);
     process.exit(1);
 });